+++ openapi_client/models/constant_lateral.py
@@ -3,7 +3,7 @@
 """
     3Di API
 
-    3Di simulation API (latest version: 3.0)   Framework release: 0.0.27   3Di core release: 2.0.3  deployed on:  09:45AM (UTC) on December 02, 2019  # noqa: E501
+    3Di simulation API (latest version: 3.0)   Framework release: 0.0.29   3Di core release: 2.0.3  deployed on:  02:21PM (UTC) on December 02, 2019  # noqa: E501
 
     The version of the OpenAPI document: 3.0
     Contact: info@nelen-schuurmans.nl
@@ -40,7 +40,6 @@ class ConstantLateral(object):
         'duration': 'int',
         'value': 'float',
         'units': 'str',
-        'epsg': 'int',
         'point': 'str',
         'connection_node': 'int',
         'state': 'str',
@@ -55,7 +54,6 @@ class ConstantLateral(object):
         'duration': 'duration',
         'value': 'value',
         'units': 'units',
-        'epsg': 'epsg',
         'point': 'point',
         'connection_node': 'connection_node',
         'state': 'state',
@@ -63,7 +61,7 @@ class ConstantLateral(object):
         'grid_id': 'grid_id'
     }
 
-    def __init__(self, url=None, simulation=None, offset=None, duration=None, value=None, units=None, epsg=None, point=None, connection_node=None, state=None, state_detail=None, grid_id=None, local_vars_configuration=None):  # noqa: E501
+    def __init__(self, url=None, simulation=None, offset=None, duration=None, value=None, units=None, point=None, connection_node=None, state=None, state_detail=None, grid_id=None, local_vars_configuration=None):  # noqa: E501
         """ConstantLateral - a model defined in OpenAPI"""  # noqa: E501
         if local_vars_configuration is None:
             local_vars_configuration = Configuration()
@@ -75,7 +73,6 @@ class ConstantLateral(object):
         self._duration = None
         self._value = None
         self._units = None
-        self._epsg = None
         self._point = None
         self._connection_node = None
         self._state = None
@@ -91,7 +88,6 @@ class ConstantLateral(object):
         self.duration = duration
         self.value = value
         self.units = units
-        self.epsg = epsg
         self.point = point
         self.connection_node = connection_node
         if state is not None:
@@ -256,33 +252,6 @@ class ConstantLateral(object):
             )
 
         self._units = units
-
-    @property
-    def epsg(self):
-        """Gets the epsg of this ConstantLateral.  # noqa: E501
-
-
-        :return: The epsg of this ConstantLateral.  # noqa: E501
-        :rtype: int
-        """
-        return self._epsg
-
-    @epsg.setter
-    def epsg(self, epsg):
-        """Sets the epsg of this ConstantLateral.
-
-
-        :param epsg: The epsg of this ConstantLateral.  # noqa: E501
-        :type: int
-        """
-        if (self.local_vars_configuration.client_side_validation and
-                epsg is not None and epsg > 2147483647):  # noqa: E501
-            raise ValueError("Invalid value for `epsg`, must be a value less than or equal to `2147483647`")  # noqa: E501
-        if (self.local_vars_configuration.client_side_validation and
-                epsg is not None and epsg < -2147483648):  # noqa: E501
-            raise ValueError("Invalid value for `epsg`, must be a value greater than or equal to `-2147483648`")  # noqa: E501
-
-        self._epsg = epsg
 
     @property
     def point(self):