+++ openapi_client/models/lateral.py
@@ -3,7 +3,7 @@
 """
     3Di API
 
-    3Di simulation API (latest version: 3.0)   Framework release: 0.0.27   3Di core release: 2.0.3  deployed on:  09:45AM (UTC) on December 02, 2019  # noqa: E501
+    3Di simulation API (latest version: 3.0)   Framework release: 0.0.29   3Di core release: 2.0.3  deployed on:  02:21PM (UTC) on December 02, 2019  # noqa: E501
 
     The version of the OpenAPI document: 3.0
     Contact: info@nelen-schuurmans.nl
@@ -42,7 +42,6 @@ class Lateral(object):
         'values': 'list[list[float]]',
         'units': 'str',
         'constant': 'bool',
-        'epsg': 'int',
         'point': 'str',
         'connection_node': 'int',
         'state': 'str',
@@ -59,7 +58,6 @@ class Lateral(object):
         'values': 'values',
         'units': 'units',
         'constant': 'constant',
-        'epsg': 'epsg',
         'point': 'point',
         'connection_node': 'connection_node',
         'state': 'state',
@@ -67,7 +65,7 @@ class Lateral(object):
         'grid_id': 'grid_id'
     }
 
-    def __init__(self, url=None, simulation=None, offset=None, duration=None, interpolate=None, values=None, units=None, constant=None, epsg=None, point=None, connection_node=None, state=None, state_detail=None, grid_id=None, local_vars_configuration=None):  # noqa: E501
+    def __init__(self, url=None, simulation=None, offset=None, duration=None, interpolate=None, values=None, units=None, constant=None, point=None, connection_node=None, state=None, state_detail=None, grid_id=None, local_vars_configuration=None):  # noqa: E501
         """Lateral - a model defined in OpenAPI"""  # noqa: E501
         if local_vars_configuration is None:
             local_vars_configuration = Configuration()
@@ -81,7 +79,6 @@ class Lateral(object):
         self._values = None
         self._units = None
         self._constant = None
-        self._epsg = None
         self._point = None
         self._connection_node = None
         self._state = None
@@ -101,7 +98,6 @@ class Lateral(object):
         self.units = units
         if constant is not None:
             self.constant = constant
-        self.epsg = epsg
         self.point = point
         self.connection_node = connection_node
         if state is not None:
@@ -308,33 +304,6 @@ class Lateral(object):
         """
 
         self._constant = constant
-
-    @property
-    def epsg(self):
-        """Gets the epsg of this Lateral.  # noqa: E501
-
-
-        :return: The epsg of this Lateral.  # noqa: E501
-        :rtype: int
-        """
-        return self._epsg
-
-    @epsg.setter
-    def epsg(self, epsg):
-        """Sets the epsg of this Lateral.
-
-
-        :param epsg: The epsg of this Lateral.  # noqa: E501
-        :type: int
-        """
-        if (self.local_vars_configuration.client_side_validation and
-                epsg is not None and epsg > 2147483647):  # noqa: E501
-            raise ValueError("Invalid value for `epsg`, must be a value less than or equal to `2147483647`")  # noqa: E501
-        if (self.local_vars_configuration.client_side_validation and
-                epsg is not None and epsg < -2147483648):  # noqa: E501
-            raise ValueError("Invalid value for `epsg`, must be a value greater than or equal to `-2147483648`")  # noqa: E501
-
-        self._epsg = epsg
 
     @property
     def point(self):